PR-E-7: extract MemberDefEmitter (property/event accessor IL emission)
Move property and event accessor IL emission out of
ReflectionMetadataEmitter into a dedicated internal sealed class
MemberDefEmitter, following the same composition pattern as
DataStructSynthesizer (E-6) and ConversionEmitter (E-5).

Methods moved (15 + 4 private helpers):

Instance property accessors:
- EmitPropertyAccessors
- EmitPropertyGetter
- EmitPropertySetter

Static property accessors:
- EmitStaticPropertyAccessors
- EmitStaticPropertyGetter
- EmitStaticPropertySetter

Instance event accessors:
- EmitEventAccessors
- EmitEventAddAccessor
- EmitEventRemoveAccessor
- EmitEventRaiseAccessor

Static event accessors:
- EmitStaticEventAccessors
- EmitStaticEventAddAccessor
- EmitStaticEventRemoveAccessor

Interface accessors (abstract, no body):
- EmitInterfacePropertyAccessors
- EmitInterfaceEventAccessors

Private helpers (used exclusively by the above):
- PropertyImplicitlyImplementsInterface
- EventImplicitlyImplementsInterface
- GetEventTypeHandle
- GetInterlockedCompareExchangeSpec

What stays:
- TypeDef and constructor emission (EmitStructTypeDef,
  EmitClassConstructor*, EmitDefaultConstructor) -> PR-E-8 TypeDefEmitter.
- BodyEmitter internals -> PR-E-11.
- Cache dictionaries (PropertyAccessorHandles, EventAccessorHandles,
  TypesWithPropertyMap) remain on MetadataTokenCache (PR-E-2).

Wiring:
- ReflectionMetadataEmitter constructs MemberDefEmitter in EmitCore
  with injected EmitContext, MetadataTokenCache, WellKnownReferences
  plus delegates for EmitFunction, EncodeTypeSymbol,
  NextParameterHandle, GetTypeReference, GetTypeHandleForMember.
- All 10 internal call sites updated:
  this.EmitX(...) -> this.memberDefEmitter.EmitX(...)
- MemberDefEmitter has no hard reference to ReflectionMetadataEmitter.

LoC delta: ReflectionMetadataEmitter.cs 15587 -> 14229 (-1358).
MemberDefEmitter.cs: 1492 lines.

Validation:
- dotnet build green (0 warnings, 0 errors).
- RefactoringBaselineTests: zero hash diffs.
- DeterministicEmitTests: green.
- Core.Tests, Interpreter.Tests, LanguageServer.Tests, Sdk.Tests:
  all green.
- Compiler.Tests pass when run individually; full-project run hits
  macOS fd exhaustion (256 soft limit) - environmental, not a
  regression. CI on Linux runs the full suite.

Next PR in sequence: E-8 TypeDefEmitter.
