+++ src/Core/CodeAnalysis/Emit/ReflectionMetadataEmitter.cs
@@ -200,6 +200,24 @@ internal sealed class ReflectionMetadataEmitter
     // ConversionEmitter / wellKnown.
     private DataStructSynthesizer dataStructSynth;
 
+    // PR-E-7: MemberDefEmitter — sibling of DataStructSynthesizer — owns
+    // IL emission for property and event accessor MethodDefs plus the
+    // PropertyDef/EventDef/PropertyMap/EventMap/MethodSemantics rows that
+    // link them to their owning TypeDef. Covers instance, static, and
+    // interface variants for both properties and events. Like
+    // DataStructSynthesizer, every method moved here was a top-level
+    // private on this emitter — none lived inside the nested BodyEmitter —
+    // so the extraction is a clean Option-A move. The shared
+    // PropertyAccessorHandles / EventAccessorHandles / TypesWithPropertyMap
+    // collections stay on MetadataTokenCache; this component reads/writes
+    // them through `cache.*`. The four PR-E-7 private helpers
+    // (PropertyImplicitlyImplementsInterface, EventImplicitlyImplementsInterface,
+    // GetEventTypeHandle, GetInterlockedCompareExchangeSpec) all moved with
+    // it because their callers were exclusively the 15 accessor-emit methods.
+    // Not readonly for the same EmitCore-ordering reason as
+    // ConversionEmitter / dataStructSynth / wellKnown.
+    private MemberDefEmitter memberDefEmitter;
+
     private ReflectionMetadataEmitter(BoundProgram program, ReferenceResolver references, string assemblyName, bool metadataOnly)
     {
         this.emitCtx = new EmitContext(program, references, assemblyName, metadataOnly);
@@ -388,6 +406,23 @@ internal sealed class ReflectionMetadataEmitter
             this.GetElementTypeToken,
             this.GetTypeReference,
             this.NextParameterHandle);
+
+        // PR-E-7: MemberDefEmitter wires up after DataStructSynthesizer.
+        // It depends on the same EmitContext/MetadataTokenCache/WellKnownReferences
+        // trio and threads delegate callbacks for the five root-emitter
+        // helpers it uses (EmitFunction, EncodeTypeSymbol, NextParameterHandle,
+        // GetTypeReference, GetTypeHandleForMember) — same composition pattern
+        // as DataStructSynthesizer / ConversionEmitter / SlotPlanner. No hard
+        // back-reference to this emitter.
+        this.memberDefEmitter = new MemberDefEmitter(
+            this.emitCtx,
+            this.cache,
+            this.wellKnown,
+            this.EmitFunction,
+            this.EncodeTypeSymbol,
+            this.NextParameterHandle,
+            this.GetTypeReference,
+            this.GetTypeHandleForMember);
 
         // Pre-assign FieldDefinitionHandles for user struct fields. Struct
         // TypeDefs are emitted between <Module> and the per-package <Program>
@@ -1280,10 +1315,10 @@ internal sealed class ReflectionMetadataEmitter
             }
 
             // Issue #248: emit abstract accessor MethodDefs + PropertyDef rows for interface properties.
-            this.EmitInterfacePropertyAccessors(i);
+            this.memberDefEmitter.EmitInterfacePropertyAccessors(i);
 
             // ADR-0052: emit abstract accessor MethodDefs + EventDef rows for interface events.
-            this.EmitInterfaceEventAccessors(i);
+            this.memberDefEmitter.EmitInterfaceEventAccessors(i);
         }
 
         // B2. Non-SM class ctors + instance methods.
@@ -1352,10 +1387,10 @@ internal sealed class ReflectionMetadataEmitter
             }
 
             // ADR-0051 Phase 6: emit property accessor methods for classes.
-            this.EmitPropertyAccessors(c);
+            this.memberDefEmitter.EmitPropertyAccessors(c);
 
             // ADR-0052: emit event accessor methods for classes.
-            this.EmitEventAccessors(c);
+            this.memberDefEmitter.EmitEventAccessors(c);
 
             // ADR-0053: emit static methods for classes.
             if (!c.StaticMethods.IsDefaultOrEmpty)
@@ -1371,10 +1406,10 @@ internal sealed class ReflectionMetadataEmitter
             }
 
             // Issue #263: emit static property accessor methods for classes.
-            this.EmitStaticPropertyAccessors(c);
+            this.memberDefEmitter.EmitStaticPropertyAccessors(c);
 
             // Issue #263: emit static event accessor methods for classes.
-            this.EmitStaticEventAccessors(c);
+            this.memberDefEmitter.EmitStaticEventAccessors(c);
 
             // Issue #262: emit .cctor for classes with static field initializers.
             if (this.cache.CctorHandles.ContainsKey(c))
@@ -1415,10 +1450,10 @@ internal sealed class ReflectionMetadataEmitter
             }
 
             // ADR-0051 Phase 6: emit property accessor methods for structs.
-            this.EmitPropertyAccessors(s);
+            this.memberDefEmitter.EmitPropertyAccessors(s);
 
             // ADR-0052: emit event accessor methods for structs.
-            this.EmitEventAccessors(s);
+            this.memberDefEmitter.EmitEventAccessors(s);
 
             // ADR-0053: emit static methods for structs.
             if (!s.StaticMethods.IsDefaultOrEmpty)
@@ -1434,10 +1469,10 @@ internal sealed class ReflectionMetadataEmitter
             }
 
             // Issue #263: emit static property accessor methods for structs.
-            this.EmitStaticPropertyAccessors(s);
+            this.memberDefEmitter.EmitStaticPropertyAccessors(s);
 
             // Issue #263: emit static event accessor methods for structs.
-            this.EmitStaticEventAccessors(s);
+            this.memberDefEmitter.EmitStaticEventAccessors(s);
 
             // Issue #262: emit .cctor for structs with static field initializers.
             if (this.cache.CctorHandles.ContainsKey(s))
@@ -2222,725 +2257,6 @@ internal sealed class ReflectionMetadataEmitter
     }
 
     /// <summary>
-    /// ADR-0051 Phase 6: emits accessor MethodDefs, PropertyDef rows, PropertyMap,
-    /// and MethodSemantics rows for all properties declared on a type.
-    /// Called during Phase B (MethodDef emission) after the type's regular methods.
-    /// </summary>
-    private void EmitPropertyAccessors(StructSymbol structSym)
-    {
-        if (structSym.Properties.IsDefaultOrEmpty)
-        {
-            return;
-        }
-
-        if (!this.cache.StructTypeDefs.TryGetValue(structSym, out var typeDefHandle))
-        {
-            return;
-        }
-
-        PropertyDefinitionHandle firstPropDef = default;
-        foreach (var prop in structSym.Properties)
-        {
-            if (!this.cache.PropertyAccessorHandles.TryGetValue(prop, out var accessorHandles))
-            {
-                continue;
-            }
-
-            // Emit getter MethodDef.
-            MethodDefinitionHandle? emittedGetter = null;
-            if (prop.HasGetter && accessorHandles.Getter.HasValue)
-            {
-                emittedGetter = this.EmitPropertyGetter(structSym, prop);
-            }
-
-            // Emit setter MethodDef.
-            MethodDefinitionHandle? emittedSetter = null;
-            if (prop.HasSetter && accessorHandles.Setter.HasValue)
-            {
-                emittedSetter = this.EmitPropertySetter(structSym, prop);
-            }
-
-            // Emit PropertyDef row.
-            var propertySignature = new BlobBuilder();
-            new BlobEncoder(propertySignature)
-                .PropertySignature(isInstanceProperty: true)
-                .Parameters(0, returnType => this.EncodeTypeSymbol(returnType.Type(), prop.Type), parameters => { });
-
-            var propDef = this.emitCtx.Metadata.AddProperty(
-                attributes: PropertyAttributes.None,
-                name: this.emitCtx.Metadata.GetOrAddString(prop.Name),
-                signature: this.emitCtx.Metadata.GetOrAddBlob(propertySignature));
-
-            if (firstPropDef.IsNil)
-            {
-                firstPropDef = propDef;
-            }
-
-            // MethodSemantics rows linking accessor MethodDefs to the PropertyDef.
-            if (emittedGetter.HasValue)
-            {
-                this.emitCtx.Metadata.AddMethodSemantics(propDef, MethodSemanticsAttributes.Getter, emittedGetter.Value);
-            }
-
-            if (emittedSetter.HasValue)
-            {
-                this.emitCtx.Metadata.AddMethodSemantics(propDef, MethodSemanticsAttributes.Setter, emittedSetter.Value);
-            }
-        }
-
-        // PropertyMap row: links the TypeDef to its first PropertyDef.
-        if (!firstPropDef.IsNil)
-        {
-            this.emitCtx.Metadata.AddPropertyMap(typeDefHandle, firstPropDef);
-            this.cache.TypesWithPropertyMap.Add(typeDefHandle);
-        }
-    }
-
-    /// <summary>
-    /// ADR-0051 Phase 6: emits a getter accessor MethodDef (get_PropertyName).
-    /// For auto-properties: ldarg.0, ldfld backing, ret.
-    /// For computed properties: emits the bound getter body IL.
-    /// </summary>
-    private MethodDefinitionHandle EmitPropertyGetter(StructSymbol structSym, PropertySymbol prop)
-    {
-        int bodyOffset = -1;
-        if (!this.emitCtx.MetadataOnly)
-        {
-            if (prop.IsAutoProperty && prop.BackingField != null
-                && this.cache.StructFieldDefs.TryGetValue(prop.BackingField, out var backingHandle))
-            {
-                var il = new InstructionEncoder(new BlobBuilder());
-                il.LoadArgument(0);
-                il.OpCode(ILOpCode.Ldfld);
-                il.Token(backingHandle);
-                il.OpCode(ILOpCode.Ret);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-            else if (prop.GetterSymbol != null && this.emitCtx.Program.Functions.TryGetValue(prop.GetterSymbol, out var getterBody))
-            {
-                // Computed property with bound body: emit using EmitFunction infrastructure.
-                var handle = this.EmitFunction(prop.GetterSymbol, getterBody, isEntryPoint: false);
-                return handle;
-            }
-            else
-            {
-                // Fallback: throw new NotImplementedException().
-                var il = new InstructionEncoder(new BlobBuilder());
-                var nieCtor = this.wellKnown.GetNotImplementedExceptionCtor();
-                il.OpCode(ILOpCode.Newobj);
-                il.Token(nieCtor);
-                il.OpCode(ILOpCode.Throw);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-        }
-
-        var sigBlob = new BlobBuilder();
-        new BlobEncoder(sigBlob).MethodSignature(isInstanceMethod: true)
-            .Parameters(0, r => this.EncodeTypeSymbol(r.Type(), prop.Type), _ => { });
-
-        var methodAttrs = MethodAttributes.Public | MethodAttributes.SpecialName | MethodAttributes.HideBySig;
-        if (prop.IsVirtual)
-        {
-            methodAttrs |= MethodAttributes.Virtual | MethodAttributes.NewSlot;
-        }
-        else if (prop.IsOverride)
-        {
-            methodAttrs |= MethodAttributes.Virtual;
-        }
-        else if (this.PropertyImplicitlyImplementsInterface(structSym, prop))
-        {
-            // Issue #248: implicit interface implementation requires Virtual | NewSlot.
-            methodAttrs |= MethodAttributes.Virtual | MethodAttributes.NewSlot;
-        }
-
-        return this.emitCtx.Metadata.AddMethodDefinition(
-            attributes: methodAttrs,
-            implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-            name: this.emitCtx.Metadata.GetOrAddString($"get_{prop.Name}"),
-            signature: this.emitCtx.Metadata.GetOrAddBlob(sigBlob),
-            bodyOffset: bodyOffset,
-            parameterList: this.NextParameterHandle());
-    }
-
-    /// <summary>
-    /// ADR-0051 Phase 6: emits a setter accessor MethodDef (set_PropertyName).
-    /// For auto-properties: ldarg.0, ldarg.1, stfld backing, ret.
-    /// For computed properties: emits the bound setter body IL.
-    /// </summary>
-    private MethodDefinitionHandle EmitPropertySetter(StructSymbol structSym, PropertySymbol prop)
-    {
-        int bodyOffset = -1;
-        if (!this.emitCtx.MetadataOnly)
-        {
-            if (prop.IsAutoProperty && prop.BackingField != null
-                && this.cache.StructFieldDefs.TryGetValue(prop.BackingField, out var backingHandle))
-            {
-                var il = new InstructionEncoder(new BlobBuilder());
-                il.LoadArgument(0);
-                il.LoadArgument(1);
-                il.OpCode(ILOpCode.Stfld);
-                il.Token(backingHandle);
-                il.OpCode(ILOpCode.Ret);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-            else if (prop.SetterSymbol != null && this.emitCtx.Program.Functions.TryGetValue(prop.SetterSymbol, out var setterBody))
-            {
-                // Computed property with bound body: emit using EmitFunction infrastructure.
-                var handle = this.EmitFunction(prop.SetterSymbol, setterBody, isEntryPoint: false);
-                return handle;
-            }
-            else
-            {
-                // Fallback: throw new NotImplementedException().
-                var il = new InstructionEncoder(new BlobBuilder());
-                var nieCtor = this.wellKnown.GetNotImplementedExceptionCtor();
-                il.OpCode(ILOpCode.Newobj);
-                il.Token(nieCtor);
-                il.OpCode(ILOpCode.Throw);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-        }
-
-        var sigBlob = new BlobBuilder();
-        new BlobEncoder(sigBlob).MethodSignature(isInstanceMethod: true)
-            .Parameters(1, r => r.Void(), ps => this.EncodeTypeSymbol(ps.AddParameter().Type(), prop.Type));
-
-        var methodAttrs = MethodAttributes.Public | MethodAttributes.SpecialName | MethodAttributes.HideBySig;
-        if (prop.IsVirtual)
-        {
-            methodAttrs |= MethodAttributes.Virtual | MethodAttributes.NewSlot;
-        }
-        else if (prop.IsOverride)
-        {
-            methodAttrs |= MethodAttributes.Virtual;
-        }
-        else if (this.PropertyImplicitlyImplementsInterface(structSym, prop))
-        {
-            // Issue #248: implicit interface implementation requires Virtual | NewSlot.
-            methodAttrs |= MethodAttributes.Virtual | MethodAttributes.NewSlot;
-        }
-
-        // Emit a Parameter row for "value" so the setter has a named parameter.
-        var firstParamHandle = this.NextParameterHandle();
-        this.emitCtx.Metadata.AddParameter(
-            attributes: ParameterAttributes.None,
-            name: this.emitCtx.Metadata.GetOrAddString(prop.SetterParameterName ?? "value"),
-            sequenceNumber: 1);
-
-        return this.emitCtx.Metadata.AddMethodDefinition(
-            attributes: methodAttrs,
-            implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-            name: this.emitCtx.Metadata.GetOrAddString($"set_{prop.Name}"),
-            signature: this.emitCtx.Metadata.GetOrAddBlob(sigBlob),
-            bodyOffset: bodyOffset,
-            parameterList: firstParamHandle);
-    }
-
-    /// <summary>
-    /// Issue #263: emits accessor MethodDefs, PropertyDef rows, PropertyMap,
-    /// and MethodSemantics rows for static properties declared in a shared block.
-    /// </summary>
-    private void EmitStaticPropertyAccessors(StructSymbol structSym)
-    {
-        if (structSym.StaticProperties.IsDefaultOrEmpty)
-        {
-            return;
-        }
-
-        if (!this.cache.StructTypeDefs.TryGetValue(structSym, out var typeDefHandle))
-        {
-            return;
-        }
-
-        PropertyDefinitionHandle firstPropDef = default;
-        foreach (var prop in structSym.StaticProperties)
-        {
-            if (!this.cache.PropertyAccessorHandles.TryGetValue(prop, out var accessorHandles))
-            {
-                continue;
-            }
-
-            // Emit getter MethodDef.
-            MethodDefinitionHandle? emittedGetter = null;
-            if (prop.HasGetter && accessorHandles.Getter.HasValue)
-            {
-                emittedGetter = this.EmitStaticPropertyGetter(structSym, prop);
-            }
-
-            // Emit setter MethodDef.
-            MethodDefinitionHandle? emittedSetter = null;
-            if (prop.HasSetter && accessorHandles.Setter.HasValue)
-            {
-                emittedSetter = this.EmitStaticPropertySetter(structSym, prop);
-            }
-
-            // Emit PropertyDef row.
-            var propertySignature = new BlobBuilder();
-            new BlobEncoder(propertySignature)
-                .PropertySignature(isInstanceProperty: false)
-                .Parameters(0, returnType => this.EncodeTypeSymbol(returnType.Type(), prop.Type), parameters => { });
-
-            var propDef = this.emitCtx.Metadata.AddProperty(
-                attributes: PropertyAttributes.None,
-                name: this.emitCtx.Metadata.GetOrAddString(prop.Name),
-                signature: this.emitCtx.Metadata.GetOrAddBlob(propertySignature));
-
-            if (firstPropDef.IsNil)
-            {
-                firstPropDef = propDef;
-            }
-
-            // MethodSemantics rows linking accessor MethodDefs to the PropertyDef.
-            if (emittedGetter.HasValue)
-            {
-                this.emitCtx.Metadata.AddMethodSemantics(propDef, MethodSemanticsAttributes.Getter, emittedGetter.Value);
-            }
-
-            if (emittedSetter.HasValue)
-            {
-                this.emitCtx.Metadata.AddMethodSemantics(propDef, MethodSemanticsAttributes.Setter, emittedSetter.Value);
-            }
-        }
-
-        // PropertyMap row: links the TypeDef to its first PropertyDef.
-        // Issue #418 (P1-7): only add a PropertyMap here if the instance-property
-        // emission path (EmitPropertyAccessors) didn't already add one. Using
-        // structSym.Properties.IsDefaultOrEmpty was incorrect — instance properties
-        // may be declared but all skipped during emission (e.g., computed property
-        // whose getter symbol has no entry in program.Functions), leaving no
-        // PropertyMap. Without this row the static PropertyDef rows would be
-        // orphaned and violate ECMA-335 §II.22.35.
-        if (!firstPropDef.IsNil && !this.cache.TypesWithPropertyMap.Contains(typeDefHandle))
-        {
-            this.emitCtx.Metadata.AddPropertyMap(typeDefHandle, firstPropDef);
-            this.cache.TypesWithPropertyMap.Add(typeDefHandle);
-        }
-    }
-
-    /// <summary>
-    /// Issue #263: emits a static getter accessor MethodDef (get_PropertyName).
-    /// </summary>
-    private MethodDefinitionHandle EmitStaticPropertyGetter(StructSymbol structSym, PropertySymbol prop)
-    {
-        int bodyOffset = -1;
-        if (!this.emitCtx.MetadataOnly)
-        {
-            if (prop.IsAutoProperty && prop.BackingField != null
-                && this.cache.StructFieldDefs.TryGetValue(prop.BackingField, out var backingHandle))
-            {
-                var il = new InstructionEncoder(new BlobBuilder());
-                il.OpCode(ILOpCode.Ldsfld);
-                il.Token(backingHandle);
-                il.OpCode(ILOpCode.Ret);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-            else if (prop.GetterSymbol != null && this.emitCtx.Program.Functions.TryGetValue(prop.GetterSymbol, out var getterBody))
-            {
-                var handle = this.EmitFunction(prop.GetterSymbol, getterBody, isEntryPoint: false);
-                return handle;
-            }
-            else
-            {
-                var il = new InstructionEncoder(new BlobBuilder());
-                var nieCtor = this.wellKnown.GetNotImplementedExceptionCtor();
-                il.OpCode(ILOpCode.Newobj);
-                il.Token(nieCtor);
-                il.OpCode(ILOpCode.Throw);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-        }
-
-        var sigBlob = new BlobBuilder();
-        new BlobEncoder(sigBlob).MethodSignature(isInstanceMethod: false)
-            .Parameters(0, r => this.EncodeTypeSymbol(r.Type(), prop.Type), _ => { });
-
-        var methodAttrs = MethodAttributes.Public | MethodAttributes.SpecialName | MethodAttributes.HideBySig | MethodAttributes.Static;
-
-        return this.emitCtx.Metadata.AddMethodDefinition(
-            attributes: methodAttrs,
-            implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-            name: this.emitCtx.Metadata.GetOrAddString($"get_{prop.Name}"),
-            signature: this.emitCtx.Metadata.GetOrAddBlob(sigBlob),
-            bodyOffset: bodyOffset,
-            parameterList: this.NextParameterHandle());
-    }
-
-    /// <summary>
-    /// Issue #263: emits a static setter accessor MethodDef (set_PropertyName).
-    /// </summary>
-    private MethodDefinitionHandle EmitStaticPropertySetter(StructSymbol structSym, PropertySymbol prop)
-    {
-        int bodyOffset = -1;
-        if (!this.emitCtx.MetadataOnly)
-        {
-            if (prop.IsAutoProperty && prop.BackingField != null
-                && this.cache.StructFieldDefs.TryGetValue(prop.BackingField, out var backingHandle))
-            {
-                var il = new InstructionEncoder(new BlobBuilder());
-                il.LoadArgument(0);
-                il.OpCode(ILOpCode.Stsfld);
-                il.Token(backingHandle);
-                il.OpCode(ILOpCode.Ret);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-            else if (prop.SetterSymbol != null && this.emitCtx.Program.Functions.TryGetValue(prop.SetterSymbol, out var setterBody))
-            {
-                var handle = this.EmitFunction(prop.SetterSymbol, setterBody, isEntryPoint: false);
-                return handle;
-            }
-            else
-            {
-                var il = new InstructionEncoder(new BlobBuilder());
-                var nieCtor = this.wellKnown.GetNotImplementedExceptionCtor();
-                il.OpCode(ILOpCode.Newobj);
-                il.Token(nieCtor);
-                il.OpCode(ILOpCode.Throw);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-        }
-
-        var sigBlob = new BlobBuilder();
-        new BlobEncoder(sigBlob).MethodSignature(isInstanceMethod: false)
-            .Parameters(1, r => r.Void(), ps => this.EncodeTypeSymbol(ps.AddParameter().Type(), prop.Type));
-
-        var methodAttrs = MethodAttributes.Public | MethodAttributes.SpecialName | MethodAttributes.HideBySig | MethodAttributes.Static;
-
-        var firstParamHandle = this.NextParameterHandle();
-        this.emitCtx.Metadata.AddParameter(
-            attributes: ParameterAttributes.None,
-            name: this.emitCtx.Metadata.GetOrAddString(prop.SetterParameterName ?? "value"),
-            sequenceNumber: 1);
-
-        return this.emitCtx.Metadata.AddMethodDefinition(
-            attributes: methodAttrs,
-            implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-            name: this.emitCtx.Metadata.GetOrAddString($"set_{prop.Name}"),
-            signature: this.emitCtx.Metadata.GetOrAddBlob(sigBlob),
-            bodyOffset: bodyOffset,
-            parameterList: firstParamHandle);
-    }
-
-    /// <summary>
-    /// Issue #263: emits add/remove accessor MethodDefs, EventDef rows, EventMap,
-    /// and MethodSemantics rows for static events declared in a shared block.
-    /// </summary>
-    private void EmitStaticEventAccessors(StructSymbol structSym)
-    {
-        if (structSym.StaticEvents.IsDefaultOrEmpty)
-        {
-            return;
-        }
-
-        if (!this.cache.StructTypeDefs.TryGetValue(structSym, out var typeDefHandle))
-        {
-            return;
-        }
-
-        EventDefinitionHandle firstEventDef = default;
-        foreach (var ev in structSym.StaticEvents)
-        {
-            if (!this.cache.EventAccessorHandles.TryGetValue(ev, out var accessorHandles))
-            {
-                continue;
-            }
-
-            // Emit add_X MethodDef.
-            var addMethod = this.EmitStaticEventAddAccessor(structSym, ev);
-
-            // Emit remove_X MethodDef.
-            var removeMethod = this.EmitStaticEventRemoveAccessor(structSym, ev);
-
-            // Issue #257: emit raise_X MethodDef if present.
-            MethodDefinitionHandle? raiseMethod = null;
-            if (ev.RaiseMethodSymbol != null)
-            {
-                raiseMethod = this.EmitEventRaiseAccessor(structSym, ev, isStatic: true);
-            }
-
-            // Emit EventDef row.
-            var eventTypeHandle = this.GetEventTypeHandle(ev.Type);
-
-            var eventDef = this.emitCtx.Metadata.AddEvent(
-                attributes: EventAttributes.None,
-                name: this.emitCtx.Metadata.GetOrAddString(ev.Name),
-                type: eventTypeHandle);
-
-            if (firstEventDef.IsNil)
-            {
-                firstEventDef = eventDef;
-            }
-
-            // MethodSemantics rows linking accessor MethodDefs to the EventDef.
-            this.emitCtx.Metadata.AddMethodSemantics(eventDef, MethodSemanticsAttributes.Adder, addMethod);
-            this.emitCtx.Metadata.AddMethodSemantics(eventDef, MethodSemanticsAttributes.Remover, removeMethod);
-            if (raiseMethod.HasValue)
-            {
-                this.emitCtx.Metadata.AddMethodSemantics(eventDef, MethodSemanticsAttributes.Raiser, raiseMethod.Value);
-            }
-        }
-
-        // EventMap row: links the TypeDef to its first EventDef.
-        // Only add if no instance events already created an EventMap for this type.
-        if (!firstEventDef.IsNil && structSym.Events.IsDefaultOrEmpty)
-        {
-            this.emitCtx.Metadata.AddEventMap(typeDefHandle, firstEventDef);
-        }
-    }
-
-    /// <summary>
-    /// Issue #263: emits a static add_X accessor MethodDef for a static event.
-    /// </summary>
-    private MethodDefinitionHandle EmitStaticEventAddAccessor(StructSymbol structSym, EventSymbol ev)
-    {
-        int bodyOffset = -1;
-        if (!this.emitCtx.MetadataOnly)
-        {
-            if (ev.IsFieldLike && ev.BackingField != null
-                && this.cache.StructFieldDefs.TryGetValue(ev.BackingField, out var backingHandle))
-            {
-                // Issue #256: thread-safe CAS loop using Interlocked.CompareExchange<T>.
-                var il = new InstructionEncoder(new BlobBuilder(), new ControlFlowBuilder());
-                var eventTypeHandle = this.GetEventTypeHandle(ev.Type);
-
-                // ldsfld backingField; stloc.0
-                il.OpCode(ILOpCode.Ldsfld);
-                il.Token(backingHandle);
-                il.OpCode(ILOpCode.Stloc_0);
-
-                // loop_start:
-                var loopStart = il.DefineLabel();
-                il.MarkLabel(loopStart);
-
-                // ldloc.0; stloc.1
-                il.OpCode(ILOpCode.Ldloc_0);
-                il.OpCode(ILOpCode.Stloc_1);
-
-                // ldloc.1; ldarg.0; call Delegate.Combine; castclass T; stloc.2
-                il.OpCode(ILOpCode.Ldloc_1);
-                il.LoadArgument(0);
-                il.OpCode(ILOpCode.Call);
-                il.Token(this.wellKnown.GetDelegateCombineRef());
-                il.OpCode(ILOpCode.Castclass);
-                il.Token(eventTypeHandle);
-                il.OpCode(ILOpCode.Stloc_2);
-
-                // ldsflda backingField; ldloc.2; ldloc.1
-                il.OpCode(ILOpCode.Ldsflda);
-                il.Token(backingHandle);
-                il.OpCode(ILOpCode.Ldloc_2);
-                il.OpCode(ILOpCode.Ldloc_1);
-
-                // call Interlocked.CompareExchange<T>; stloc.0
-                il.OpCode(ILOpCode.Call);
-                il.Token(this.GetInterlockedCompareExchangeSpec(ev.Type));
-                il.OpCode(ILOpCode.Stloc_0);
-
-                // ldloc.0; ldloc.1; bne.un.s loop_start
-                il.OpCode(ILOpCode.Ldloc_0);
-                il.OpCode(ILOpCode.Ldloc_1);
-                il.Branch(ILOpCode.Bne_un_s, loopStart);
-
-                il.OpCode(ILOpCode.Ret);
-
-                var localsSigBlob = new BlobBuilder();
-                var localsEncoder = new BlobEncoder(localsSigBlob).LocalVariableSignature(3);
-                this.EncodeTypeSymbol(localsEncoder.AddVariable().Type(), ev.Type);
-                this.EncodeTypeSymbol(localsEncoder.AddVariable().Type(), ev.Type);
-                this.EncodeTypeSymbol(localsEncoder.AddVariable().Type(), ev.Type);
-                var localsSignature = this.emitCtx.Metadata.AddStandaloneSignature(this.emitCtx.Metadata.GetOrAddBlob(localsSigBlob));
-
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il, maxStack: 3, localVariablesSignature: localsSignature);
-            }
-            else if (ev.AddMethodSymbol != null && this.emitCtx.Program.Functions.TryGetValue(ev.AddMethodSymbol, out var addBody))
-            {
-                var handle = this.EmitFunction(ev.AddMethodSymbol, addBody, isEntryPoint: false);
-                return handle;
-            }
-            else
-            {
-                var il = new InstructionEncoder(new BlobBuilder());
-                var nieCtor = this.wellKnown.GetNotImplementedExceptionCtor();
-                il.OpCode(ILOpCode.Newobj);
-                il.Token(nieCtor);
-                il.OpCode(ILOpCode.Throw);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-        }
-
-        var sigBlob = new BlobBuilder();
-        new BlobEncoder(sigBlob).MethodSignature(isInstanceMethod: false)
-            .Parameters(1, r => r.Void(), ps => this.EncodeTypeSymbol(ps.AddParameter().Type(), ev.Type));
-
-        var methodAttrs = MethodAttributes.Public | MethodAttributes.SpecialName | MethodAttributes.HideBySig | MethodAttributes.Static;
-
-        var firstParamHandle = this.NextParameterHandle();
-        this.emitCtx.Metadata.AddParameter(
-            attributes: ParameterAttributes.None,
-            name: this.emitCtx.Metadata.GetOrAddString("value"),
-            sequenceNumber: 1);
-
-        return this.emitCtx.Metadata.AddMethodDefinition(
-            attributes: methodAttrs,
-            implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-            name: this.emitCtx.Metadata.GetOrAddString($"add_{ev.Name}"),
-            signature: this.emitCtx.Metadata.GetOrAddBlob(sigBlob),
-            bodyOffset: bodyOffset,
-            parameterList: firstParamHandle);
-    }
-
-    /// <summary>
-    /// Issue #263: emits a static remove_X accessor MethodDef for a static event.
-    /// </summary>
-    private MethodDefinitionHandle EmitStaticEventRemoveAccessor(StructSymbol structSym, EventSymbol ev)
-    {
-        int bodyOffset = -1;
-        if (!this.emitCtx.MetadataOnly)
-        {
-            if (ev.IsFieldLike && ev.BackingField != null
-                && this.cache.StructFieldDefs.TryGetValue(ev.BackingField, out var backingHandle))
-            {
-                // Issue #256: thread-safe CAS loop using Interlocked.CompareExchange<T>.
-                var il = new InstructionEncoder(new BlobBuilder(), new ControlFlowBuilder());
-                var eventTypeHandle = this.GetEventTypeHandle(ev.Type);
-
-                // ldsfld backingField; stloc.0
-                il.OpCode(ILOpCode.Ldsfld);
-                il.Token(backingHandle);
-                il.OpCode(ILOpCode.Stloc_0);
-
-                // loop_start:
-                var loopStart = il.DefineLabel();
-                il.MarkLabel(loopStart);
-
-                // ldloc.0; stloc.1
-                il.OpCode(ILOpCode.Ldloc_0);
-                il.OpCode(ILOpCode.Stloc_1);
-
-                // ldloc.1; ldarg.0; call Delegate.Remove; castclass T; stloc.2
-                il.OpCode(ILOpCode.Ldloc_1);
-                il.LoadArgument(0);
-                il.OpCode(ILOpCode.Call);
-                il.Token(this.wellKnown.GetDelegateRemoveRef());
-                il.OpCode(ILOpCode.Castclass);
-                il.Token(eventTypeHandle);
-                il.OpCode(ILOpCode.Stloc_2);
-
-                // ldsflda backingField; ldloc.2; ldloc.1
-                il.OpCode(ILOpCode.Ldsflda);
-                il.Token(backingHandle);
-                il.OpCode(ILOpCode.Ldloc_2);
-                il.OpCode(ILOpCode.Ldloc_1);
-
-                // call Interlocked.CompareExchange<T>; stloc.0
-                il.OpCode(ILOpCode.Call);
-                il.Token(this.GetInterlockedCompareExchangeSpec(ev.Type));
-                il.OpCode(ILOpCode.Stloc_0);
-
-                // ldloc.0; ldloc.1; bne.un.s loop_start
-                il.OpCode(ILOpCode.Ldloc_0);
-                il.OpCode(ILOpCode.Ldloc_1);
-                il.Branch(ILOpCode.Bne_un_s, loopStart);
-
-                il.OpCode(ILOpCode.Ret);
-
-                var localsSigBlob = new BlobBuilder();
-                var localsEncoder = new BlobEncoder(localsSigBlob).LocalVariableSignature(3);
-                this.EncodeTypeSymbol(localsEncoder.AddVariable().Type(), ev.Type);
-                this.EncodeTypeSymbol(localsEncoder.AddVariable().Type(), ev.Type);
-                this.EncodeTypeSymbol(localsEncoder.AddVariable().Type(), ev.Type);
-                var localsSignature = this.emitCtx.Metadata.AddStandaloneSignature(this.emitCtx.Metadata.GetOrAddBlob(localsSigBlob));
-
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il, maxStack: 3, localVariablesSignature: localsSignature);
-            }
-            else if (ev.RemoveMethodSymbol != null && this.emitCtx.Program.Functions.TryGetValue(ev.RemoveMethodSymbol, out var removeBody))
-            {
-                var handle = this.EmitFunction(ev.RemoveMethodSymbol, removeBody, isEntryPoint: false);
-                return handle;
-            }
-            else
-            {
-                var il = new InstructionEncoder(new BlobBuilder());
-                var nieCtor = this.wellKnown.GetNotImplementedExceptionCtor();
-                il.OpCode(ILOpCode.Newobj);
-                il.Token(nieCtor);
-                il.OpCode(ILOpCode.Throw);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-        }
-
-        var sigBlob = new BlobBuilder();
-        new BlobEncoder(sigBlob).MethodSignature(isInstanceMethod: false)
-            .Parameters(1, r => r.Void(), ps => this.EncodeTypeSymbol(ps.AddParameter().Type(), ev.Type));
-
-        var methodAttrs = MethodAttributes.Public | MethodAttributes.SpecialName | MethodAttributes.HideBySig | MethodAttributes.Static;
-
-        var firstParamHandle = this.NextParameterHandle();
-        this.emitCtx.Metadata.AddParameter(
-            attributes: ParameterAttributes.None,
-            name: this.emitCtx.Metadata.GetOrAddString("value"),
-            sequenceNumber: 1);
-
-        return this.emitCtx.Metadata.AddMethodDefinition(
-            attributes: methodAttrs,
-            implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-            name: this.emitCtx.Metadata.GetOrAddString($"remove_{ev.Name}"),
-            signature: this.emitCtx.Metadata.GetOrAddBlob(sigBlob),
-            bodyOffset: bodyOffset,
-            parameterList: firstParamHandle);
-    }
-
-    /// <summary>
-    /// Issue #248: determines whether a property on a class/struct implicitly implements
-    /// an interface property (same name and type on any implemented interface).
-    /// </summary>
-    private bool PropertyImplicitlyImplementsInterface(StructSymbol structSym, PropertySymbol prop)
-    {
-        if (!structSym.Interfaces.IsDefaultOrEmpty)
-        {
-            foreach (var iface in structSym.Interfaces)
-            {
-                if (iface.Properties.IsDefaultOrEmpty)
-                {
-                    continue;
-                }
-
-                foreach (var ifaceProp in iface.Properties)
-                {
-                    if (ifaceProp.Name == prop.Name)
-                    {
-                        return true;
-                    }
-                }
-            }
-        }
-
-        // Issue #525: imported CLR interfaces from the base-type clause also
-        // trigger the implicit-implementation virtual-slot promotion.
-        if (!structSym.ImplementedClrInterfaces.IsDefaultOrEmpty)
-        {
-            var propClr = prop.Type?.ClrType;
-            foreach (var ifaceSym in structSym.ImplementedClrInterfaces)
-            {
-                var clrIface = ifaceSym?.ClrType;
-                if (clrIface == null)
-                {
-                    continue;
-                }
-
-                foreach (var clrProp in clrIface.GetProperties(System.Reflection.BindingFlags.Public | System.Reflection.BindingFlags.Instance))
-                {
-                    if (clrProp.Name == prop.Name && (propClr == null || ClrTypeUtilities.AreSame(propClr, clrProp.PropertyType)))
-                    {
-                        return true;
-                    }
-                }
-            }
-        }
-
-        return false;
-    }
-
-    /// <summary>
     /// Issue #409: determines whether a value-type instance method must keep
     /// virtual method attributes because it participates in CLR vtable dispatch.
     /// </summary>
@@ -3061,460 +2377,6 @@ internal sealed class ReflectionMetadataEmitter
 
     private static ImmutableArray<ParameterSymbol> CallableParameters(FunctionSymbol method)
         => method.ExplicitReceiverParameter == null ? method.Parameters : method.Parameters.RemoveAt(0);
-
-    /// <summary>
-    /// ADR-0052: determines whether an event on a class/struct implicitly implements
-    /// an interface event (same name on any implemented interface).
-    /// </summary>
-    private bool EventImplicitlyImplementsInterface(StructSymbol structSym, EventSymbol ev)
-    {
-        if (structSym.Interfaces.IsDefaultOrEmpty)
-        {
-            return false;
-        }
-
-        foreach (var iface in structSym.Interfaces)
-        {
-            if (iface.Events.IsDefaultOrEmpty)
-            {
-                continue;
-            }
-
-            foreach (var ifaceEvent in iface.Events)
-            {
-                if (ifaceEvent.Name == ev.Name)
-                {
-                    return true;
-                }
-            }
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    /// ADR-0052: emits add/remove accessor MethodDefs, EventDef rows, EventMap,
-    /// and MethodSemantics rows for all events declared on a type.
-    /// </summary>
-    private void EmitEventAccessors(StructSymbol structSym)
-    {
-        if (structSym.Events.IsDefaultOrEmpty)
-        {
-            return;
-        }
-
-        if (!this.cache.StructTypeDefs.TryGetValue(structSym, out var typeDefHandle))
-        {
-            return;
-        }
-
-        EventDefinitionHandle firstEventDef = default;
-        foreach (var ev in structSym.Events)
-        {
-            if (!this.cache.EventAccessorHandles.TryGetValue(ev, out var accessorHandles))
-            {
-                continue;
-            }
-
-            // Emit add_X MethodDef.
-            var addMethod = this.EmitEventAddAccessor(structSym, ev);
-
-            // Emit remove_X MethodDef.
-            var removeMethod = this.EmitEventRemoveAccessor(structSym, ev);
-
-            // Issue #257: emit raise_X MethodDef if present.
-            MethodDefinitionHandle? raiseMethod = null;
-            if (ev.RaiseMethodSymbol != null)
-            {
-                raiseMethod = this.EmitEventRaiseAccessor(structSym, ev, isStatic: false);
-            }
-
-            // Emit EventDef row.
-            var eventTypeHandle = this.GetEventTypeHandle(ev.Type);
-
-            var eventDef = this.emitCtx.Metadata.AddEvent(
-                attributes: EventAttributes.None,
-                name: this.emitCtx.Metadata.GetOrAddString(ev.Name),
-                type: eventTypeHandle);
-
-            if (firstEventDef.IsNil)
-            {
-                firstEventDef = eventDef;
-            }
-
-            // MethodSemantics rows linking accessor MethodDefs to the EventDef.
-            this.emitCtx.Metadata.AddMethodSemantics(eventDef, MethodSemanticsAttributes.Adder, addMethod);
-            this.emitCtx.Metadata.AddMethodSemantics(eventDef, MethodSemanticsAttributes.Remover, removeMethod);
-            if (raiseMethod.HasValue)
-            {
-                this.emitCtx.Metadata.AddMethodSemantics(eventDef, MethodSemanticsAttributes.Raiser, raiseMethod.Value);
-            }
-        }
-
-        // EventMap row: links the TypeDef to its first EventDef.
-        if (!firstEventDef.IsNil)
-        {
-            this.emitCtx.Metadata.AddEventMap(typeDefHandle, firstEventDef);
-        }
-    }
-
-    /// <summary>
-    /// ADR-0052: resolves the EntityHandle for the event handler type used in the EventDef row.
-    /// </summary>
-    private EntityHandle GetEventTypeHandle(TypeSymbol type)
-    {
-        if (type is FunctionTypeSymbol fnType)
-        {
-            var clrType = fnType.ClrType;
-            if (clrType != null)
-            {
-                return this.GetTypeHandleForMember(clrType);
-            }
-        }
-
-        if (type.ClrType != null)
-        {
-            return this.GetTypeHandleForMember(type.ClrType);
-        }
-
-        if (type is StructSymbol structSym && this.cache.StructTypeDefs.TryGetValue(structSym, out var td))
-        {
-            return td;
-        }
-
-        if (type is InterfaceSymbol ifaceSym && this.cache.InterfaceTypeDefs.TryGetValue(ifaceSym, out var ifaceDef))
-        {
-            return ifaceDef;
-        }
-
-        // Fallback: encode as System.Delegate.
-        return this.GetTypeReference(typeof(System.Delegate));
-    }
-
-    /// <summary>
-    /// ADR-0052: emits the add_X accessor MethodDef for an event.
-    /// </summary>
-    private MethodDefinitionHandle EmitEventAddAccessor(StructSymbol structSym, EventSymbol ev)
-    {
-        int bodyOffset = -1;
-        if (!this.emitCtx.MetadataOnly)
-        {
-            if (ev.IsFieldLike && ev.BackingField != null
-                && this.cache.StructFieldDefs.TryGetValue(ev.BackingField, out var backingHandle))
-            {
-                // Issue #256: thread-safe CAS loop using Interlocked.CompareExchange<T>.
-                var il = new InstructionEncoder(new BlobBuilder(), new ControlFlowBuilder());
-                var eventTypeHandle = this.GetEventTypeHandle(ev.Type);
-
-                // ldarg.0; ldfld backingField; stloc.0
-                il.LoadArgument(0);
-                il.OpCode(ILOpCode.Ldfld);
-                il.Token(backingHandle);
-                il.OpCode(ILOpCode.Stloc_0);
-
-                // loop_start:
-                var loopStart = il.DefineLabel();
-                il.MarkLabel(loopStart);
-
-                // ldloc.0; stloc.1
-                il.OpCode(ILOpCode.Ldloc_0);
-                il.OpCode(ILOpCode.Stloc_1);
-
-                // ldloc.1; ldarg.1; call Delegate.Combine; castclass T; stloc.2
-                il.OpCode(ILOpCode.Ldloc_1);
-                il.LoadArgument(1);
-                il.OpCode(ILOpCode.Call);
-                il.Token(this.wellKnown.GetDelegateCombineRef());
-                il.OpCode(ILOpCode.Castclass);
-                il.Token(eventTypeHandle);
-                il.OpCode(ILOpCode.Stloc_2);
-
-                // ldarg.0; ldflda backingField; ldloc.2; ldloc.1
-                il.LoadArgument(0);
-                il.OpCode(ILOpCode.Ldflda);
-                il.Token(backingHandle);
-                il.OpCode(ILOpCode.Ldloc_2);
-                il.OpCode(ILOpCode.Ldloc_1);
-
-                // call Interlocked.CompareExchange<T>; stloc.0
-                il.OpCode(ILOpCode.Call);
-                il.Token(this.GetInterlockedCompareExchangeSpec(ev.Type));
-                il.OpCode(ILOpCode.Stloc_0);
-
-                // ldloc.0; ldloc.1; bne.un.s loop_start
-                il.OpCode(ILOpCode.Ldloc_0);
-                il.OpCode(ILOpCode.Ldloc_1);
-                il.Branch(ILOpCode.Bne_un_s, loopStart);
-
-                il.OpCode(ILOpCode.Ret);
-
-                var localsSigBlob = new BlobBuilder();
-                var localsEncoder = new BlobEncoder(localsSigBlob).LocalVariableSignature(3);
-                this.EncodeTypeSymbol(localsEncoder.AddVariable().Type(), ev.Type);
-                this.EncodeTypeSymbol(localsEncoder.AddVariable().Type(), ev.Type);
-                this.EncodeTypeSymbol(localsEncoder.AddVariable().Type(), ev.Type);
-                var localsSignature = this.emitCtx.Metadata.AddStandaloneSignature(this.emitCtx.Metadata.GetOrAddBlob(localsSigBlob));
-
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il, maxStack: 3, localVariablesSignature: localsSignature);
-            }
-            else if (ev.AddMethodSymbol != null && this.emitCtx.Program.Functions.TryGetValue(ev.AddMethodSymbol, out var addBody))
-            {
-                // Explicit accessor with bound body: emit using EmitFunction infrastructure.
-                var handle = this.EmitFunction(ev.AddMethodSymbol, addBody, isEntryPoint: false);
-                return handle;
-            }
-            else
-            {
-                // Fallback: throw new NotImplementedException().
-                var il = new InstructionEncoder(new BlobBuilder());
-                var nieCtor = this.wellKnown.GetNotImplementedExceptionCtor();
-                il.OpCode(ILOpCode.Newobj);
-                il.Token(nieCtor);
-                il.OpCode(ILOpCode.Throw);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-        }
-
-        var sigBlob = new BlobBuilder();
-        new BlobEncoder(sigBlob).MethodSignature(isInstanceMethod: true)
-            .Parameters(1, r => r.Void(), ps => this.EncodeTypeSymbol(ps.AddParameter().Type(), ev.Type));
-
-        var methodAttrs = MethodAttributes.Public | MethodAttributes.SpecialName | MethodAttributes.HideBySig;
-        if (ev.IsVirtual)
-        {
-            methodAttrs |= MethodAttributes.Virtual | MethodAttributes.NewSlot;
-        }
-        else if (ev.IsOverride)
-        {
-            methodAttrs |= MethodAttributes.Virtual;
-        }
-        else if (this.EventImplicitlyImplementsInterface(structSym, ev))
-        {
-            methodAttrs |= MethodAttributes.Virtual | MethodAttributes.NewSlot;
-        }
-
-        var firstParamHandle = this.NextParameterHandle();
-        this.emitCtx.Metadata.AddParameter(
-            attributes: ParameterAttributes.None,
-            name: this.emitCtx.Metadata.GetOrAddString("value"),
-            sequenceNumber: 1);
-
-        return this.emitCtx.Metadata.AddMethodDefinition(
-            attributes: methodAttrs,
-            implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-            name: this.emitCtx.Metadata.GetOrAddString($"add_{ev.Name}"),
-            signature: this.emitCtx.Metadata.GetOrAddBlob(sigBlob),
-            bodyOffset: bodyOffset,
-            parameterList: firstParamHandle);
-    }
-
-    /// <summary>
-    /// ADR-0052: emits the remove_X accessor MethodDef for an event.
-    /// </summary>
-    private MethodDefinitionHandle EmitEventRemoveAccessor(StructSymbol structSym, EventSymbol ev)
-    {
-        int bodyOffset = -1;
-        if (!this.emitCtx.MetadataOnly)
-        {
-            if (ev.IsFieldLike && ev.BackingField != null
-                && this.cache.StructFieldDefs.TryGetValue(ev.BackingField, out var backingHandle))
-            {
-                // Issue #256: thread-safe CAS loop using Interlocked.CompareExchange<T>.
-                var il = new InstructionEncoder(new BlobBuilder(), new ControlFlowBuilder());
-                var eventTypeHandle = this.GetEventTypeHandle(ev.Type);
-
-                // ldarg.0; ldfld backingField; stloc.0
-                il.LoadArgument(0);
-                il.OpCode(ILOpCode.Ldfld);
-                il.Token(backingHandle);
-                il.OpCode(ILOpCode.Stloc_0);
-
-                // loop_start:
-                var loopStart = il.DefineLabel();
-                il.MarkLabel(loopStart);
-
-                // ldloc.0; stloc.1
-                il.OpCode(ILOpCode.Ldloc_0);
-                il.OpCode(ILOpCode.Stloc_1);
-
-                // ldloc.1; ldarg.1; call Delegate.Remove; castclass T; stloc.2
-                il.OpCode(ILOpCode.Ldloc_1);
-                il.LoadArgument(1);
-                il.OpCode(ILOpCode.Call);
-                il.Token(this.wellKnown.GetDelegateRemoveRef());
-                il.OpCode(ILOpCode.Castclass);
-                il.Token(eventTypeHandle);
-                il.OpCode(ILOpCode.Stloc_2);
-
-                // ldarg.0; ldflda backingField; ldloc.2; ldloc.1
-                il.LoadArgument(0);
-                il.OpCode(ILOpCode.Ldflda);
-                il.Token(backingHandle);
-                il.OpCode(ILOpCode.Ldloc_2);
-                il.OpCode(ILOpCode.Ldloc_1);
-
-                // call Interlocked.CompareExchange<T>; stloc.0
-                il.OpCode(ILOpCode.Call);
-                il.Token(this.GetInterlockedCompareExchangeSpec(ev.Type));
-                il.OpCode(ILOpCode.Stloc_0);
-
-                // ldloc.0; ldloc.1; bne.un.s loop_start
-                il.OpCode(ILOpCode.Ldloc_0);
-                il.OpCode(ILOpCode.Ldloc_1);
-                il.Branch(ILOpCode.Bne_un_s, loopStart);
-
-                il.OpCode(ILOpCode.Ret);
-
-                var localsSigBlob = new BlobBuilder();
-                var localsEncoder = new BlobEncoder(localsSigBlob).LocalVariableSignature(3);
-                this.EncodeTypeSymbol(localsEncoder.AddVariable().Type(), ev.Type);
-                this.EncodeTypeSymbol(localsEncoder.AddVariable().Type(), ev.Type);
-                this.EncodeTypeSymbol(localsEncoder.AddVariable().Type(), ev.Type);
-                var localsSignature = this.emitCtx.Metadata.AddStandaloneSignature(this.emitCtx.Metadata.GetOrAddBlob(localsSigBlob));
-
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il, maxStack: 3, localVariablesSignature: localsSignature);
-            }
-            else if (ev.RemoveMethodSymbol != null && this.emitCtx.Program.Functions.TryGetValue(ev.RemoveMethodSymbol, out var removeBody))
-            {
-                // Explicit accessor with bound body: emit using EmitFunction infrastructure.
-                var handle = this.EmitFunction(ev.RemoveMethodSymbol, removeBody, isEntryPoint: false);
-                return handle;
-            }
-            else
-            {
-                // Fallback: throw new NotImplementedException().
-                var il = new InstructionEncoder(new BlobBuilder());
-                var nieCtor = this.wellKnown.GetNotImplementedExceptionCtor();
-                il.OpCode(ILOpCode.Newobj);
-                il.Token(nieCtor);
-                il.OpCode(ILOpCode.Throw);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-        }
-
-        var sigBlob = new BlobBuilder();
-        new BlobEncoder(sigBlob).MethodSignature(isInstanceMethod: true)
-            .Parameters(1, r => r.Void(), ps => this.EncodeTypeSymbol(ps.AddParameter().Type(), ev.Type));
-
-        var methodAttrs = MethodAttributes.Public | MethodAttributes.SpecialName | MethodAttributes.HideBySig;
-        if (ev.IsVirtual)
-        {
-            methodAttrs |= MethodAttributes.Virtual | MethodAttributes.NewSlot;
-        }
-        else if (ev.IsOverride)
-        {
-            methodAttrs |= MethodAttributes.Virtual;
-        }
-        else if (this.EventImplicitlyImplementsInterface(structSym, ev))
-        {
-            methodAttrs |= MethodAttributes.Virtual | MethodAttributes.NewSlot;
-        }
-
-        var firstParamHandle = this.NextParameterHandle();
-        this.emitCtx.Metadata.AddParameter(
-            attributes: ParameterAttributes.None,
-            name: this.emitCtx.Metadata.GetOrAddString("value"),
-            sequenceNumber: 1);
-
-        return this.emitCtx.Metadata.AddMethodDefinition(
-            attributes: methodAttrs,
-            implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-            name: this.emitCtx.Metadata.GetOrAddString($"remove_{ev.Name}"),
-            signature: this.emitCtx.Metadata.GetOrAddBlob(sigBlob),
-            bodyOffset: bodyOffset,
-            parameterList: firstParamHandle);
-    }
-
-    /// <summary>
-    /// Issue #257: emits the raise_X accessor MethodDef for an event.
-    /// The raise accessor body is always user-provided (explicit).
-    /// </summary>
-    private MethodDefinitionHandle EmitEventRaiseAccessor(StructSymbol structSym, EventSymbol ev, bool isStatic)
-    {
-        int bodyOffset = -1;
-        if (!this.emitCtx.MetadataOnly)
-        {
-            if (ev.RaiseMethodSymbol != null && this.emitCtx.Program.Functions.TryGetValue(ev.RaiseMethodSymbol, out var raiseBody))
-            {
-                var handle = this.EmitFunction(ev.RaiseMethodSymbol, raiseBody, isEntryPoint: false);
-                return handle;
-            }
-            else
-            {
-                // Fallback: throw new NotImplementedException().
-                var il = new InstructionEncoder(new BlobBuilder());
-                var nieCtor = this.wellKnown.GetNotImplementedExceptionCtor();
-                il.OpCode(ILOpCode.Newobj);
-                il.Token(nieCtor);
-                il.OpCode(ILOpCode.Throw);
-                bodyOffset = this.emitCtx.MethodBodyStream.AddMethodBody(il);
-            }
-        }
-
-        // Build signature: parameters match the handler type's parameters, return void.
-        int paramCount = 0;
-        if (ev.Type is FunctionTypeSymbol fnType)
-        {
-            paramCount = fnType.ParameterTypes.Length;
-        }
-
-        var sigBlob = new BlobBuilder();
-        new BlobEncoder(sigBlob).MethodSignature(isInstanceMethod: !isStatic)
-            .Parameters(paramCount, r => r.Void(), ps =>
-            {
-                if (ev.Type is FunctionTypeSymbol fn)
-                {
-                    foreach (var pt in fn.ParameterTypes)
-                    {
-                        this.EncodeTypeSymbol(ps.AddParameter().Type(), pt);
-                    }
-                }
-            });
-
-        var methodAttrs = MethodAttributes.Public | MethodAttributes.SpecialName | MethodAttributes.HideBySig;
-        if (isStatic)
-        {
-            methodAttrs |= MethodAttributes.Static;
-        }
-        else if (ev.IsVirtual)
-        {
-            methodAttrs |= MethodAttributes.Virtual | MethodAttributes.NewSlot;
-        }
-        else if (ev.IsOverride)
-        {
-            methodAttrs |= MethodAttributes.Virtual;
-        }
-
-        var firstParamHandle = this.NextParameterHandle();
-        for (int i = 0; i < paramCount; i++)
-        {
-            this.emitCtx.Metadata.AddParameter(
-                attributes: ParameterAttributes.None,
-                name: this.emitCtx.Metadata.GetOrAddString($"arg{i}"),
-                sequenceNumber: i + 1);
-        }
-
-        return this.emitCtx.Metadata.AddMethodDefinition(
-            attributes: methodAttrs,
-            implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-            name: this.emitCtx.Metadata.GetOrAddString($"raise_{ev.Name}"),
-            signature: this.emitCtx.Metadata.GetOrAddBlob(sigBlob),
-            bodyOffset: bodyOffset,
-            parameterList: firstParamHandle);
-    }
-
-    /// <summary>
-    /// Issue #256: produces a MethodSpec for Interlocked.CompareExchange&lt;EventType&gt;.
-    /// </summary>
-    private EntityHandle GetInterlockedCompareExchangeSpec(TypeSymbol eventType)
-    {
-        var openRef = this.wellKnown.GetInterlockedCompareExchangeOpenRef();
-        var sigBlob = new BlobBuilder();
-        var argsEncoder = new BlobEncoder(sigBlob).MethodSpecificationSignature(1);
-        var typeEncoder = argsEncoder.AddArgument();
-        this.EncodeTypeSymbol(typeEncoder, eventType);
-        return this.emitCtx.Metadata.AddMethodSpecification(openRef, this.emitCtx.Metadata.GetOrAddBlob(sigBlob));
-    }
 
     /// <summary>
     /// Emits a nested-private TypeDef for a state-machine type.  Same as
@@ -4814,226 +3676,6 @@ internal sealed class ReflectionMetadataEmitter
         // ADR-0047 §3: user annotations targeting the delegate type land on
         // the TypeDef row (same as struct/interface/enum).
         this.EmitUserAttributes(delegateTypeDef, delegateSym, AttributeTargetKind.Type);
-    }
-
-    /// <summary>
-    /// Issue #248: emits abstract accessor MethodDefs, PropertyDef rows, PropertyMap,
-    /// and MethodSemantics rows for all properties declared on an interface.
-    /// </summary>
-    private void EmitInterfacePropertyAccessors(InterfaceSymbol ifaceSym)
-    {
-        if (ifaceSym.Properties.IsDefaultOrEmpty)
-        {
-            return;
-        }
-
-        if (!this.cache.InterfaceTypeDefs.TryGetValue(ifaceSym, out var typeDefHandle))
-        {
-            return;
-        }
-
-        PropertyDefinitionHandle firstPropDef = default;
-        foreach (var prop in ifaceSym.Properties)
-        {
-            if (!this.cache.PropertyAccessorHandles.TryGetValue(prop, out var accessorHandles))
-            {
-                continue;
-            }
-
-            // Emit abstract getter MethodDef.
-            MethodDefinitionHandle? emittedGetter = null;
-            if (prop.HasGetter && accessorHandles.Getter.HasValue)
-            {
-                var sigBlob = new BlobBuilder();
-                new BlobEncoder(sigBlob).MethodSignature(isInstanceMethod: true)
-                    .Parameters(0, r => this.EncodeTypeSymbol(r.Type(), prop.Type), _ => { });
-
-                var attrs = MethodAttributes.Public | MethodAttributes.HideBySig
-                    | MethodAttributes.Virtual | MethodAttributes.Abstract
-                    | MethodAttributes.NewSlot | MethodAttributes.SpecialName;
-                emittedGetter = this.emitCtx.Metadata.AddMethodDefinition(
-                    attributes: attrs,
-                    implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-                    name: this.emitCtx.Metadata.GetOrAddString($"get_{prop.Name}"),
-                    signature: this.emitCtx.Metadata.GetOrAddBlob(sigBlob),
-                    bodyOffset: -1,
-                    parameterList: this.NextParameterHandle());
-            }
-
-            // Emit abstract setter MethodDef.
-            MethodDefinitionHandle? emittedSetter = null;
-            if (prop.HasSetter && accessorHandles.Setter.HasValue)
-            {
-                var sigBlob = new BlobBuilder();
-                new BlobEncoder(sigBlob).MethodSignature(isInstanceMethod: true)
-                    .Parameters(1, r => r.Void(), ps =>
-                    {
-                        this.EncodeTypeSymbol(ps.AddParameter().Type(), prop.Type);
-                    });
-
-                var attrs = MethodAttributes.Public | MethodAttributes.HideBySig
-                    | MethodAttributes.Virtual | MethodAttributes.Abstract
-                    | MethodAttributes.NewSlot | MethodAttributes.SpecialName;
-                emittedSetter = this.emitCtx.Metadata.AddMethodDefinition(
-                    attributes: attrs,
-                    implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-                    name: this.emitCtx.Metadata.GetOrAddString($"set_{prop.Name}"),
-                    signature: this.emitCtx.Metadata.GetOrAddBlob(sigBlob),
-                    bodyOffset: -1,
-                    parameterList: this.NextParameterHandle());
-            }
-
-            // Emit PropertyDef row.
-            var propertySignature = new BlobBuilder();
-            new BlobEncoder(propertySignature)
-                .PropertySignature(isInstanceProperty: true)
-                .Parameters(0, returnType => this.EncodeTypeSymbol(returnType.Type(), prop.Type), parameters => { });
-
-            var propDef = this.emitCtx.Metadata.AddProperty(
-                attributes: PropertyAttributes.None,
-                name: this.emitCtx.Metadata.GetOrAddString(prop.Name),
-                signature: this.emitCtx.Metadata.GetOrAddBlob(propertySignature));
-
-            if (firstPropDef.IsNil)
-            {
-                firstPropDef = propDef;
-            }
-
-            // MethodSemantics rows linking accessor MethodDefs to the PropertyDef.
-            if (emittedGetter.HasValue)
-            {
-                this.emitCtx.Metadata.AddMethodSemantics(propDef, MethodSemanticsAttributes.Getter, emittedGetter.Value);
-            }
-
-            if (emittedSetter.HasValue)
-            {
-                this.emitCtx.Metadata.AddMethodSemantics(propDef, MethodSemanticsAttributes.Setter, emittedSetter.Value);
-            }
-        }
-
-        // PropertyMap row: links the TypeDef to its first PropertyDef.
-        if (!firstPropDef.IsNil)
-        {
-            this.emitCtx.Metadata.AddPropertyMap(typeDefHandle, firstPropDef);
-            this.cache.TypesWithPropertyMap.Add(typeDefHandle);
-        }
-    }
-
-    /// <summary>
-    /// ADR-0052: emits abstract add/remove accessor MethodDefs, EventDef rows, EventMap,
-    /// and MethodSemantics rows for all events declared on an interface.
-    /// </summary>
-    private void EmitInterfaceEventAccessors(InterfaceSymbol ifaceSym)
-    {
-        if (ifaceSym.Events.IsDefaultOrEmpty)
-        {
-            return;
-        }
-
-        if (!this.cache.InterfaceTypeDefs.TryGetValue(ifaceSym, out var typeDefHandle))
-        {
-            return;
-        }
-
-        EventDefinitionHandle firstEventDef = default;
-        foreach (var ev in ifaceSym.Events)
-        {
-            if (!this.cache.EventAccessorHandles.TryGetValue(ev, out var accessorHandles))
-            {
-                continue;
-            }
-
-            // Emit abstract add_X MethodDef.
-            var addSigBlob = new BlobBuilder();
-            new BlobEncoder(addSigBlob).MethodSignature(isInstanceMethod: true)
-                .Parameters(1, r => r.Void(), ps => this.EncodeTypeSymbol(ps.AddParameter().Type(), ev.Type));
-
-            var addAttrs = MethodAttributes.Public | MethodAttributes.HideBySig
-                | MethodAttributes.Virtual | MethodAttributes.Abstract
-                | MethodAttributes.NewSlot | MethodAttributes.SpecialName;
-            var emittedAdd = this.emitCtx.Metadata.AddMethodDefinition(
-                attributes: addAttrs,
-                implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-                name: this.emitCtx.Metadata.GetOrAddString($"add_{ev.Name}"),
-                signature: this.emitCtx.Metadata.GetOrAddBlob(addSigBlob),
-                bodyOffset: -1,
-                parameterList: this.NextParameterHandle());
-
-            // Emit abstract remove_X MethodDef.
-            var removeSigBlob = new BlobBuilder();
-            new BlobEncoder(removeSigBlob).MethodSignature(isInstanceMethod: true)
-                .Parameters(1, r => r.Void(), ps => this.EncodeTypeSymbol(ps.AddParameter().Type(), ev.Type));
-
-            var removeAttrs = MethodAttributes.Public | MethodAttributes.HideBySig
-                | MethodAttributes.Virtual | MethodAttributes.Abstract
-                | MethodAttributes.NewSlot | MethodAttributes.SpecialName;
-            var emittedRemove = this.emitCtx.Metadata.AddMethodDefinition(
-                attributes: removeAttrs,
-                implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-                name: this.emitCtx.Metadata.GetOrAddString($"remove_{ev.Name}"),
-                signature: this.emitCtx.Metadata.GetOrAddBlob(removeSigBlob),
-                bodyOffset: -1,
-                parameterList: this.NextParameterHandle());
-
-            // Emit EventDef row.
-            var eventTypeHandle = this.GetEventTypeHandle(ev.Type);
-
-            var eventDef = this.emitCtx.Metadata.AddEvent(
-                attributes: EventAttributes.None,
-                name: this.emitCtx.Metadata.GetOrAddString(ev.Name),
-                type: eventTypeHandle);
-
-            if (firstEventDef.IsNil)
-            {
-                firstEventDef = eventDef;
-            }
-
-            // MethodSemantics rows linking accessor MethodDefs to the EventDef.
-            this.emitCtx.Metadata.AddMethodSemantics(eventDef, MethodSemanticsAttributes.Adder, emittedAdd);
-            this.emitCtx.Metadata.AddMethodSemantics(eventDef, MethodSemanticsAttributes.Remover, emittedRemove);
-
-            // Issue #257: emit abstract raise_X MethodDef if present.
-            if (ev.RaiseMethodSymbol != null)
-            {
-                int raiseParamCount = 0;
-                if (ev.Type is FunctionTypeSymbol fnType)
-                {
-                    raiseParamCount = fnType.ParameterTypes.Length;
-                }
-
-                var raiseSigBlob = new BlobBuilder();
-                new BlobEncoder(raiseSigBlob).MethodSignature(isInstanceMethod: true)
-                    .Parameters(raiseParamCount, r => r.Void(), ps =>
-                    {
-                        if (ev.Type is FunctionTypeSymbol fn)
-                        {
-                            foreach (var pt in fn.ParameterTypes)
-                            {
-                                this.EncodeTypeSymbol(ps.AddParameter().Type(), pt);
-                            }
-                        }
-                    });
-
-                var raiseAttrs = MethodAttributes.Public | MethodAttributes.HideBySig
-                    | MethodAttributes.Virtual | MethodAttributes.Abstract
-                    | MethodAttributes.NewSlot | MethodAttributes.SpecialName;
-                var emittedRaise = this.emitCtx.Metadata.AddMethodDefinition(
-                    attributes: raiseAttrs,
-                    implAttributes: MethodImplAttributes.IL | MethodImplAttributes.Managed,
-                    name: this.emitCtx.Metadata.GetOrAddString($"raise_{ev.Name}"),
-                    signature: this.emitCtx.Metadata.GetOrAddBlob(raiseSigBlob),
-                    bodyOffset: -1,
-                    parameterList: this.NextParameterHandle());
-
-                this.emitCtx.Metadata.AddMethodSemantics(eventDef, MethodSemanticsAttributes.Raiser, emittedRaise);
-            }
-        }
-
-        // EventMap row: links the TypeDef to its first EventDef.
-        if (!firstEventDef.IsNil)
-        {
-            this.emitCtx.Metadata.AddEventMap(typeDefHandle, firstEventDef);
-        }
     }
 
     /// <summary>
